+++ comicbook-microservice/src/main/java/comicbook/microsservice/comicbookmicroservice/repository/AutorRepository.java
@@ -3,5 +3,5 @@ package comicbook.microsservice.comicbookmicroservice.repository;
 import comicbook.microsservice.comicbookmicroservice.model.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutorRepository extends JpaRepository<Autor, Integer> {
+public interface AutorRepository extends JpaRepository<Autor, Long> {
 }
