+++ comicbook-microservice/src/main/java/comicbook/microsservice/comicbookmicroservice/repository/IzdavacRepository.java
@@ -3,5 +3,5 @@ package comicbook.microsservice.comicbookmicroservice.repository;
 import comicbook.microsservice.comicbookmicroservice.model.Izdavac;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IzdavacRepository extends JpaRepository<Izdavac, Integer> {
+public interface IzdavacRepository extends JpaRepository<Izdavac, Long> {
 }
